refact: Prepare the bot for production

+++ src/actionHandlers/commands/clearChat.ts
@@ -4,16 +4,19 @@ import type { Command } from './types';
 
 export const clearChat: Command = (interaction) => {
   const { member, channel, options } = interaction;
-  const amount = options.getNumber('amount') || 0;
+
+  if (!(member.permissions as Readonly<PermissionsBitField>).has('Administrator')) return;
+
+  const amount = options.getInteger('amount');
 
   if (amount > 100)
     return interaction.reply({ content: "Can't clear more than 100 messages at a time", ephemeral: true });
   if (amount < 1) return interaction.reply({ content: "Can't clear less than 1 message", ephemeral: true });
 
-  (member.permissions as Readonly<PermissionsBitField>).has('Administrator') &&
-    channel
-      .bulkDelete(amount)
-      .then(() => interaction.reply({ content: `Deleted ${amount} messages`, ephemeral: true }));
+  channel
+    .bulkDelete(amount)
+    .then(() => interaction.reply({ content: `Deleted ${amount} messages`, ephemeral: true }))
+    .catch(console.log);
 };
 
 clearChat.create = {
@@ -24,7 +27,7 @@ clearChat.create = {
       name: 'amount',
       description: 'The amount wanted to be cleared',
       required: true,
-      type: ApplicationCommandOptionType.Number,
+      type: ApplicationCommandOptionType.Integer,
     },
   ],
 };
